+++ ChangeLog.md
@@ -5,6 +5,7 @@
 - Allow lower case player names
 - Don't show players without points in game report
 - Removed unnecessary goalie statistics from game report
+- Fixed goalie average calculation
 
 ## 0.7.0
 - Shortened views to fit within 25 lines